+++ R/general_glm_main_function.R
@@ -46,7 +46,7 @@
 #' 
 #' - `weighted` Applies a weighted likelihood function where each observation is weighted by the inverse of its subsampling probability.
 #' 
-#' - `logOddsCorrection` This lieklihood is available only for logistic regression model (i.e., when family is binomial or quasibinomial). It uses a conditional likelihood, where each element of the likelihood represents the probability of \eqn{Y=1}, given that this subsample was drawn.
+#' - `logOddsCorrection` This likelihood is available only for logistic regression model (i.e., when family is binomial or quasibinomial). It uses a conditional likelihood, where each element of the likelihood represents the probability of \eqn{Y=1}, given that this subsample was drawn.
 #'   
 #' @param contrasts An optional list. It specifies how categorical variables are represented in the design matrix. For example, `contrasts = list(v1 = 'contr.treatment', v2 = 'contr.sum')`.
 #' 
@@ -55,9 +55,9 @@
 #' - `alpha` \eqn{\in [0,1]} is the mixture weight of the user-assigned subsampling
 #' probability and uniform subsampling probability. The actual subsample
 #' probability is \eqn{\pi = (1-\alpha)\pi^{opt} + \alpha \pi^{uni}}. This protects the estimator from extreme small
-#' subsampling probability. The default value is 0.
+#' subsampling probability. The default value is 0.01.
 #' 
-#' - `b` is a positive number which is used to constaint the poisson subsampling probability. `b` close to 0 results in subsampling probabilities closer to uniform probability \eqn{\frac{1}{N}}. `b=2` is the default value. See relevant references for further details.
+#' - `b` is a positive number which is used to constraint the poisson subsampling probability. `b` close to 0 results in subsampling probabilities closer to uniform probability \eqn{\frac{1}{N}}. `b=2` is the default value. See relevant references for further details.
 #' 
 #' @param ... A list of parameters which will be passed to `svyglm()`. 
 #'
@@ -75,6 +75,7 @@
 #'   \item{index.ssp}{Row indices of of optimal subsample in the full dataset.}
 #'   \item{N}{The number of observations in the full dataset.}
 #'   \item{subsample.size.expect}{The expected subsample size, equals to `n.ssp` for `ssp.glm.` Note that for other functions, such as \link{ssp.relogit}, this value may differ.}
+#'   \item{comp.time}{The total time of computing subsampling probabilities, drawing subsample and fitting model on the subsample.}
 #'   \item{terms}{The terms object for the fitted model.}
 #' }
 #'
@@ -113,7 +114,8 @@
 #' sigmax  <- matrix(corr, d, d) + diag(1-corr, d)
 #' X <- MASS::mvrnorm(N, rep(0, d), sigmax)
 #' Y <- rbinom(N, 1, 1 - 1 / (1 + exp(beta0[1] + X %*% beta0[-1])))
-#' data <- as.data.frame(cbind(Y, X))
+#' colnames(X) <- paste0("X", 1:d)
+#' data <- data.frame(Y = Y, X)
 #' formula <- Y ~ .
 #' n.plt <- 500
 #' n.ssp <- 1000
@@ -152,7 +154,8 @@
 #' epsilon <- runif(N)
 #' lambda <- exp(beta0[1] + X %*% beta0[-1])
 #' Y <- rpois(N, lambda)
-#' data <- as.data.frame(cbind(Y, X))
+#' colnames(X) <- paste0("X", 1:d)
+#' data <- data.frame(Y = Y, X)
 #' formula <- Y ~ .
 #' n.plt <- 200
 #' n.ssp <- 600
@@ -182,6 +185,33 @@
 #' criterion = 'uniform')
 #' summary(Uni.subsampling.results)
 #' ##################
+#' # normal linear regression
+#' set.seed(4)
+#' N <- 1e4
+#' beta0 <- rep(-0.5, 7)
+#' d <- length(beta0) - 1
+#' corr <- 0.5
+#' sigmax  <- matrix(corr, d, d) + diag(1-corr, d)
+#' X <- MASS::mvrnorm(N, rep(0, d), sigmax)
+#' epsilon <- rnorm(N, sd = 1)
+#' Y <- beta0[1] + X %*% beta0[-1] + epsilon
+#' colnames(X) <- paste0("X", 1:d)
+#' data <- data.frame(Y = Y, X)
+#' formula <- Y ~ .
+#' n.plt <- 200
+#' n.ssp <- 1000
+#' subsampling.results <- ssp.glm(
+#' formula = formula,
+#' data = data,
+#' n.plt = n.plt,
+#' n.ssp = n.ssp,
+#' family = "gaussian",
+#' criterion = "optA", 
+#' sampling.method = 'poisson',
+#' likelihood = "weighted"
+#' )
+#' summary(subsampling.results)
+#' ##################
 #' # gamma regression
 #' set.seed(1)
 #' N <- 1e4
@@ -193,7 +223,8 @@
 #' link_function <- function(X, beta0) 1 / (beta0[1] + X %*% beta0[-1])
 #' scale <- link_function(X, beta0) / shape
 #' Y <- rgamma(N, shape = shape, scale = scale)
-#' data <- as.data.frame(cbind(Y, X))
+#' colnames(X) <- paste0("X", 1:d)
+#' data <- data.frame(Y = Y, X)
 #' formula <- Y ~ .
 #' n.plt <- 200
 #' n.ssp <- 1000
@@ -243,7 +274,7 @@ ssp.glm <- function(formula,
     colnames(X)[1] <- "Intercept"
   }
   family <- match.arg(family, c('binomial', 'poisson', 'Gamma',
-                                'quasibinomial', 'quasipoisson'))
+                                'quasibinomial', 'quasipoisson', 'gaussian'))
   criterion <- match.arg(criterion, c('optL', 'optA', 'LCC', 'uniform'))
   sampling.method <- match.arg(sampling.method, c('poisson', 'withReplacement'))
   likelihood <- match.arg(likelihood, c('logOddsCorrection', 'weighted'))
@@ -268,21 +299,23 @@ ssp.glm <- function(formula,
                  )
   
   if (criterion %in% c('optL', 'optA', 'LCC')) {
+    
+    t_start <- proc.time()[3] 
     ## pilot step
     plt.estimate.results <- pilot.estimate(inputs, ...)
     p.plt <- plt.estimate.results$p.plt
     beta.plt <- plt.estimate.results$beta.plt
     ddL.plt <- plt.estimate.results$ddL.plt
     dL.sq.plt <- plt.estimate.results$dL.sq.plt
-    ddL.plt.correction <- plt.estimate.results$ddL.plt.correction
     Lambda.plt <- plt.estimate.results$Lambda.plt
     d.psi <- plt.estimate.results$d.psi
+    cov.plt <- plt.estimate.results$cov.plt
     index.plt <- plt.estimate.results$index.plt
     
     ## subsampling step
     ssp.results <- subsampling(inputs,
                                p.plt = p.plt,
-                               ddL.plt.correction = ddL.plt.correction,
+                               ddL.plt = ddL.plt,
                                d.psi = d.psi,
                                index.plt = index.plt
                                )
@@ -316,6 +349,8 @@ ssp.glm <- function(formula,
                                    n.ssp = n.ssp,
                                    beta.plt = beta.plt,
                                    beta.ssp = beta.ssp)
+    t_end <- proc.time()[3]
+    comp.time <- t_end - t_start
     beta.cmb <- combining.results$beta.cmb
     cov.cmb <- combining.results$cov.cmb
     
@@ -325,22 +360,26 @@ ssp.glm <- function(formula,
                     coef.plt = beta.plt,
                     coef.ssp = beta.ssp,
                     coef = beta.cmb,
+                    cov.plt = cov.plt,
                     cov.ssp = cov.ssp,
                     cov = cov.cmb,
                     index.plt = index.plt,
                     index = index.ssp,
                     N = N,
                     subsample.size.expect = n.ssp,
+                    comp.time = comp.time,
                     terms = mt
                     )
     class(results) <- c("ssp.glm", "list")
     return(results)
   } else if (criterion == "uniform"){
     n.uni <- n.plt + n.ssp
+    h_N = n.uni / N #  varphi = 1
+    t_start <- proc.time()[3]
     if (sampling.method == 'withReplacement') {
       index.uni <- random.index(N, n.uni)
     } else if (sampling.method == 'poisson') {
-      index.uni <- poisson.index(N, n.uni / N)
+      index.uni <- poisson.index(N, h_N)
     }
     x.uni <- X[index.uni, ]
     y.uni = Y[index.uni]
@@ -356,14 +395,18 @@ ssp.glm <- function(formula,
     dL.sq.plt <- dL.sq(linear.predictor.uni,
                        x.uni,
                        y.uni,
-                       weights = 1 / n.uni ^ 2,
+                       weights = 1 / n.uni,
                        family = family)
+    ddL.uni.inv <-  solve(ddL.uni)
+    
     if (sampling.method == 'withReplacement') {
-      cov.uni <- solve(ddL.uni) %*% 
-        (dL.sq.plt * (1 + n.uni / N)) %*% solve(ddL.uni)
+      cov.uni <- (1 / n.uni) * ddL.uni.inv %*% 
+        (dL.sq.plt * (1 + h_N)) %*% ddL.uni.inv
     } else if (sampling.method == 'poisson') {
-      cov.uni <- solve(ddL.uni) %*% dL.sq.plt %*% solve(ddL.uni)
+      cov.uni <- (1 / n.uni) * ddL.uni.inv %*% dL.sq.plt %*% ddL.uni.inv
     }
+    t_end <- proc.time()[3]
+    comp.time <- t_end - t_start
     names(beta.uni) <- colnames(X)
     results <- list(model.call = model.call,
                     index = index.uni,
@@ -371,6 +414,7 @@ ssp.glm <- function(formula,
                     cov = cov.uni,
                     N = N,
                     subsample.size.expect = n.uni,
+                    comp.time = comp.time,
                     terms = mt
                     )
     class(results) <- c("ssp.glm", "list")