+++ NEWS.md
@@ -1,3 +1,11 @@
+# subsampling 0.2.0
+
+## Changes to `ssp.glm()`
+* Unified the representation of weights in the square score and information matrices  to improve readability.
+* Added support for `family = "gaussian"`.
+* Updated the default value of `alpha` from `0` to `0.01`.
+* Added a new output element `comp.time` that reports computation time.
+
 # subsampling 0.1.1
 
 # subsampling 0.1.0